PYIC-5597 - Refactor - Changed to ResetIpvSessionHandler.java to ResetSessionIdentity
Changed naming for new lambda and related tests and methods
