+++ lambdas/reset-session-identity/src/resources/IpvLambdaJsonLayout.json
