+++ lambdas/reset-session-identity/src/resources/log4j2.json
