+++ settings.gradle
@@ -17,7 +17,7 @@ include "lambdas",
 		"lambdas:process-journey-event",
 		"lambdas:replay-cimit-vcs",
 		"lambdas:store-identity",
-		"lambdas:reset-ipv-session",
+		"lambdas:reset-session-identity",
 		"libs:audit-service",
 		"libs:cimit-service",
 		"libs:common-services",
